+++ src/App.tsx
@@ -1,9 +1,9 @@
 import { Route, Routes, useNavigate } from "react-router-dom";
 
 import useAuth, { AuthStatus } from "./hooks/useAuth";
-import CreateStudy from "./pages/CreateStudy";
+import CreateStudyPage from "./pages/CreateStudyPage";
 import LoginPage from "./pages/LoginPage";
-import StudyList from "./pages/StudyList";
+import StudyListPage from "./pages/StudyListPage";
 import ApplicationStatusWrapper from "./pages/wrappers/ApplicationStatusWrapper";
 import EditStudyWrapper from "./pages/wrappers/EditStudyWrapper";
 import StudyDetailWrapper from "./pages/wrappers/StudyDetailWrapper";
@@ -23,7 +23,7 @@ function App() {
                 <Route
                     path="/"
                     element={
-                        <StudyList
+                        <StudyListPage
                             onCreateStudy={() => navigate("/create")}
                             onViewStudyDetail={(studyId: number) =>
                                 navigate(`/detail/${studyId}`)
@@ -33,7 +33,7 @@ function App() {
                 />
                 <Route
                     path="/create"
-                    element={<CreateStudy onBack={() => navigate("/")} />}
+                    element={<CreateStudyPage onBack={() => navigate("/")} />}
                 />
                 <Route
                     path="/detail/:studyId"
