+++ src/pages/StudyMembersPage.tsx
@@ -83,7 +83,7 @@ const studyMembersData: Record<
     },
 };
 
-export default function StudyMembers({ studyId, onBack }: StudyMembersProps) {
+export default function StudyMembersPage({ studyId, onBack }: StudyMembersProps) {
     const [members, setMembers] = useState(
         studyMembersData[studyId as keyof typeof studyMembersData]?.members ||
             []
